feat: bind the calibrated evaluator catalog as run judges (phase 155)
A judge could report exactly one score, and its name was the judge's name.
Microsoft ships eleven calibrated evaluators; none of them could be bound.
This phase opens that catalog and, in doing so, reshapes what a judge returns.

RunJudgment.Score/Reason are REMOVED in favour of a list of named JudgeScore
values (K-728). Breaking, and deliberately done now: every PublicAPI.Shipped.txt
holds only its `#nullable enable` header, so no surface has been published yet.
An empty list means "no decision"; a null Value means "measured nothing".

EvaluatorRunJudge bridges Microsoft.Extensions.AI.Evaluation's IEvaluator onto
IRunJudge, prefixing each metric as `{judge}.{metric}` (K-729). The prefix is
what keeps two evaluators reporting `Relevance` from overwriting each other,
since a score name is part of the stored uniqueness key. `:` is not usable as
the separator - RunScoreRules rejects it, which the plan had not measured.

Only a judge's HEADLINE score - named after the judge and shaped as a 0-100
Numeric - feeds the online-evaluation window and the judge-score histogram
(K-730). The calibrated evaluators grade 1-5; without this rule a healthy run
scoring 4 would sit below the low-score threshold and fire the alarm every time.

IEvalEvaluatorFactory replaces the hard-coded LocalEvaluator an eval suite is
graded with (K-731). It is a factory, not a bare IAgentEvaluator, so a custom
evaluator cannot silently discard the checks the suite declared.

Microsoft.Extensions.AI.Evaluation becomes an explicit reference on
AgentPrism.Core; it was already transitive, so the shipped graph does not grow.
The .Quality catalog is referenced by NO shipped package (K-727) - a consumer
who wants it adds it themselves, as samples/AgentPrism.Api now demonstrates.

Independent audit found two defects, both fixed and both proven by a failing
test first: the headline rule checked the name but not the Kind, and a store
failure partway through a judge's score set left a half-written set that the
retry checkpoint then froze forever (K-732).

Verified with a real OpenAI run through samples/AgentPrism.Api: the built-in
judge wrote `model` = 100 and the real RelevanceEvaluator wrote
`relevance.Relevance` = 3, with the live 0-100 average correctly staying at 100.

+++ src/AgentPrism.Core/Diagnostics/AgentPrismMetrics.cs
@@ -278,7 +278,7 @@ public sealed class AgentPrismMetrics : IDisposable
     /// <param name="agentName">Name of the scored agent.</param>
     /// <param name="tenantId">Tenant identifier.</param>
     /// <param name="score">Score, 0-100.</param>
-    public void RecordJudgeScore(string judgeName, string agentName, string tenantId, int score)
+    public void RecordJudgeScore(string judgeName, string agentName, string tenantId, double score)
         => JudgeScore.Record(
             score,
             new TagList
